+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/nathanab/beacon/compare/master...HEAD)
+- Update admin page with sortable and searchable date table
 
 ## [v2.4.1](https://github.com/nathanab/beacon/compare/v2.4.0...v2.4.1)
 - Fix focus date hook causing page reload